Update -break-insert with the -f (pending flag) to not require an additional parameter
According to [this](https://sourceware.org/gdb/onlinedocs/gdb/GDB_002fMI-Breakpoint-Commands.html#GDB_002fMI-Breakpoint-Commands) document, `-break-insert`'s `-f` flag is to tell the debugger to create a pending breakpoint if the location is not found. The additional parameter should not be necessary.

```
‘-f’
If location cannot be parsed (for example if it refers to unknown files or functions), create a pending breakpoint. Without this flag, GDB will report an error, and won’t create a breakpoint, if location cannot be parsed.

‘-d’
```