Properly assign instruction offset -data-disassemble
Existing code used to assign "offset" value for instructions in response to
-data-disassemble assuming that first instruction in sequence has offset 0.
This assumption is not correct. According to [1] "offset" is "The decimal
offset in bytes from the start of 'func-name'.", thus current code used to
produce invalid results if start address for disassembling was not the start
of the function itself. This commit fixes this by changing the offset
calculation logic, to ensure that correct value is returned.

The practical issue with invalid offset values is that Eclipse CDT
Disassembly View prints function name only for instruction which has
`offset == 0`, therefore if MI "asm_insns" response contains multiple
functions, Eclipse still wouldn't report their names in the View, except for
the first one.

[1] https://sourceware.org/gdb/onlinedocs/gdb/GDB_002fMI-Data-Manipulation.html#Result-6