Escape string in the error message
Errors messages as coming from the LLDB may contain new line characters,
which are forbidden inside of the MI reply and must be escaped in the
output.

This fixes a problem where LLDB would return the Status with a multiline
string and Eclipse would show only the first line of the message, trimming
the rest.