Fix data-info-line command result
This commit fixes an issue with -data-info-line which would return its
result without a proper token. Output should look like this:

7-data-info-line *0x4
7^done,start="0x4",end="0xc",file="file.c",line="65"

But it was looking look this:

7-data-info-line *0x4
^done,start="0x4",end="0xc",file="file.c",line="65"

As a result of a missing token, requester couldn't match the result with the
request. Apparently the fix is that token in the result record should not
beset in the constructor of the CMICmdCmdDataInfoLine class.