Add pointers formatting
Currently pointers are not formatted and their default representation is
printed in all cases. However, GDB formats pointers like numbers and it
makes a few tricks possible. For example, it fixes misunderstanding
between Eclipse CDT and the backend when it tries to compute a function
address.

In the first step, Eclipse requests an address and a size of an
expression. If size cannot be computed, Eclipse decides that an
expression is not a lvalue and then tries to request a pointer as a
number. So, this commit allows one to get a pointer as a raw number if
needed.