Disable threads state checking when the proccess is running
There is a method `ChkForStateChanges` which is invoked when state of
the process is changed including the case when the execution is resumed
after stop. However, this method contains a lot of SBThread::IsValid()
invocations which always return `false` if the process is running.
Each "invalid" thread is reported to be finished and is removed from the
list of threads. Further, when the debugger stops next time, all removed
threads are added again and each `new thread` event is reported.

Such strange behaviour confuses Eclipse a lot and forces it drop the
context for each thread every time the execution is resumed.

So, this commit adds the check whether the current state is `stepping`
or `running` before updating thread list.