Allow cmake RPATH variables to be overridden from the command line
Current CMakeLists.txt forces a specific content into the final RPATH,
in particular that it should contain an absolute path to the
installation directory. This is not always optimal, as relocatable
applications on Linux require there to be a path relative to $ORIGIN.
It is possible to ensure that $ORIGIN will be in the RPATH by using a
-Wl,rpath= with a desired value, but this is basically a hack that
bypasses cmake mechanisms. Better approach would be to modify
CMakeLists.txt to respect values that are passed from the command line
via -D option.

Default behavior is not changed by this commit.