Relax the checks for "stack-list-*" MI cmd (#46)
Relax the checks for "stack-list-*" MI cmd

The "stack-list-*" MI commands were checking whether the
thread was stopped.  However, for state of some threads
would not be stopped even when they clearly are. This was
not an issue for non-MI commands since they don't check
that the thread is stopped. This change relaxes the checks
performed by the MI commands to match the ones by the
non-MI commands.